% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_rank_test.R
\name{asymptotic_paired_rank_test}
\alias{asymptotic_paired_rank_test}
\title{Asymptomtic Paired Rank Test for two paired samples}
\usage{
asymptotic_paired_rank_test(obs1, obs2)
}
\arguments{
\item{obs1}{vector with observations of size k, where k is the number
of dependent samples.}

\item{obs2}{vector of observations corresponding to obs1}
}
\value{
A list
}
\description{
asymptotic rank test for two dependent samples that is based on
overall mid-ranks.

Null hypothesis: The two samples are exchangeble, i.e. (X_k1, X_k2) is
equal in distribution to (X_k2, X_k1) where (X_k1, X_k2)' denotes independent
and identically distributed pairs of observations from k = 1,2,...,n
subjects.

Applicability:
- Applicable for metric als well as ordinal data
- should only be used for n > 15

Serves as a replacement for the paired t-test and the Wilcoxon signed rank
test that both can only be applied to metric data.

Advantages of the paired rank test:
- invariant to monotonic transformations. Data e.g. can be logged or squared
without changing the results.

Disadvantages:
- Less efficient than the t-test if it is applicable
}
\examples{

## Beispiel aus Munzel and Brunner, An Exact Paired Rank Test, 2002
obs1 <- c(6, 3, 5, 4, 5, 3, 4, 5, 5, 4, 6, 4, 4, 5, 6)
obs2 <- c(4, 1, 3, 4, 6, 3, 3, 4, 3, 3, 5, 5, 3, 4, 5)

asymptotic_paired_rank_test(obs1, obs2)

}
