% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_rank_test.R
\name{exact_paired_rank_shift}
\alias{exact_paired_rank_shift}
\title{Shift version Paired Rank Test for two paired samples}
\usage{
exact_paired_rank_shift(obs1, obs2)
}
\arguments{
\item{obs1}{vector with observations of size k, where k is the number
of dependent samples.}

\item{obs2}{vector of observations corresponding to obs1}
}
\value{
A list
}
\description{
Exact rank test for two dependent samples that is based on
overall mid-ranks. Applicable for metric als well as ordinal data

Serves as a replacement for the paired t-test and the Wilcoxon signed rank
test that both can only be applied to metric data.

Advantages of the paired rank test:
- invariant to monotonic transformations. Data e.g. can be logged or squared
without changing the results.
}
\examples{

## Beispiel aus Munzel and Brunner, An Exact Paired Rank Test, 2002
obs1 <- c(6, 3, 5, 4, 5, 3, 4, 5, 5, 4, 6, 4, 4, 5, 6)
obs2 <- c(4, 1, 3, 4, 6, 3, 3, 4, 3, 3, 5, 5, 3, 4, 5)

exact_paired_rank_shift(obs1, obs2)

}
